# orgnote-backend/app/docs/swagger.yaml
definitions:
...
  handlers.HttpError-any:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  handlers.HttpResponse-array_models_APIToken-any:
    properties:
      data:
        items:
          $ref: '#/definitions/models.APIToken'
        type: array
      meta: {}
    type: object
  handlers.HttpResponse-handlers_OAuthRedirectData-any:
    properties:
      data:
        $ref: '#/definitions/handlers.OAuthRedirectData'
      meta: {}
    type: object
  handlers.HttpResponse-models_APIToken-any:
    properties:
      data:
        $ref: '#/definitions/models.APIToken'
      meta: {}
    type: object
...
  handlers.HttpResponse-models_UserPersonalInfo-any:
    properties:
      data:
        $ref: '#/definitions/models.UserPersonalInfo'
      meta: {}
    type: object
  handlers.OAuthRedirectData:
    properties:
      redirectUrl:
        type: string
    type: object
  handlers.SubscribeBody:
    properties:
      email:
        type: string
      token:
        type: string
    type: object
...
  models.APIToken:
    properties:
      id:
        type: string
      permission:
        type: string
      token:
        type: string
    type: object
...
  models.UserPersonalInfo:
    properties:
      active:
        type: string
      avatarUrl:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      nickName:
        type: string
      profileUrl:
        type: string
      provider:
        type: string
      spaceLimit:
        type: integer
      usedSpace:
        type: integer
    type: object
info:
  contact:
    email: artawower@protonmail.com
    name: API Support
  description: List of methods for work with Org Note.
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  termsOfService: http://swagger.io/terms/
  title: Org Note API
  version: 0.0.1
paths:
  /auth/{provider}/callback:
    get:
      consumes:
      - application/json
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Callback for OAuth
      tags:
      - auth
  /auth/{provider}/login:
    get:
      consumes:
      - application/json
      description: Entrypoint for login
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-handlers_OAuthRedirectData-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: OAuth Login
      tags:
      - auth
  /auth/account:
    delete:
      consumes:
      - application/json
      description: Delete user account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Delete user account
      tags:
      - auth
  /auth/api-tokens:
    get:
      consumes:
      - application/json
      description: Return all available API tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-array_models_APIToken-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Get API tokens
      tags:
      - auth
  /auth/logout:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Logout
      tags:
      - auth
  /auth/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe for backend features, like sync notes
      parameters:
      - description: token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.SubscribeBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Subscribe
      tags:
      - auth
  /auth/token:
    post:
      consumes:
      - application/json
      description: Create API token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-models_APIToken-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Create API token
      tags:
      - auth
  /auth/token/{tokenId}:
    delete:
      consumes:
      - application/json
      description: Delete API token
      parameters:
      - description: token id
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Delete API token
      tags:
      - auth
  /auth/verify:
    get:
      consumes:
      - application/json
      description: Return found user by provided token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HttpResponse-models_UserPersonalInfo-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HttpError-any'
      summary: Verify user
      tags:
      - auth
...
swagger: "2.0"